Wired meal controller, now saving meals to db, TODO: meal display page, wire calendar

+++ bandits-api/src/main/java/com/bandits/api/bandits_api/controllers/RecipeCollectionController.java
@@ -49,7 +49,7 @@ public class RecipeCollectionController {
     }
 
     @PostMapping("/delete")
-    public void deleteRecipe(@RequestParam Long recipeId) {
+    public void deleteRecipe(@RequestParam Integer recipeId) {
         recipeRepository.deleteById(recipeId);
     }
