+++ bandits-api/src/main/java/com/bandits/api/bandits_api/repositories/RecipeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
 
     Recipe findByUri(String uri);
